requirements.txt
behave
Appium-Python-Client
selenium
allure-behave
pylint
black
